+++ risc0/circuit/rv32im/src/execute/executor.rs
@@ -493,7 +493,7 @@ impl<'a, 'b, S: Syscall> Executor<'a, 'b, S> {
 
     #[cfg(feature = "rv32im-m3")]
     fn segment_cycles(&self) -> u32 {
-        self.block_tracker.row_count(self.pager.touched_pages())
+        self.block_tracker.row_count(self.pager.touched_pages()) as u32
     }
 
     fn inc_user_cycles(&mut self, count: usize, ecall: Option<EcallKind>) {
@@ -719,7 +719,7 @@ impl<S: Syscall> Risc0Context for Executor<'_, '_, S> {
 
         #[cfg(feature = "rv32im-m3")]
         self.block_tracker
-            .track_ecall_bigint(verify_program_size as u32);
+            .track_ecall_bigint(verify_program_size as u64);
 
         self.inc_user_cycles(verify_program_size + 1, Some(EcallKind::BigInt));
 
@@ -727,12 +727,12 @@ impl<S: Syscall> Risc0Context for Executor<'_, '_, S> {
     }
 
     #[cfg(feature = "rv32im-m3")]
-    fn on_ecall_read_end(&mut self, read_bytes: u32, read_words: u32) {
+    fn on_ecall_read_end(&mut self, read_bytes: u64, read_words: u64) {
         self.block_tracker.track_ecall_read(read_bytes, read_words);
     }
 
     #[cfg(feature = "rv32im-m3")]
-    fn on_ecall_poseidon2_end(&mut self, block_count: u32) {
+    fn on_ecall_poseidon2_end(&mut self, block_count: u64) {
         self.block_tracker.track_ecall_poseidon2(block_count);
     }
